move `watch` to `state/watch`
