+++ state/memory.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 
 	"github.com/docker/swarm-v2/api"
-	"github.com/docker/swarm-v2/watch"
+	"github.com/docker/swarm-v2/state/watch"
 	memdb "github.com/hashicorp/go-memdb"
 )
